abstracted and organized funcitons; css cleanup
